dormitory

+++ src/dao/NoticeDao.java
@@ -11,6 +11,6 @@ public interface NoticeDao {
     public void delete(int id);         //根据公告id删除一条公告
     public Notice find(int id);         //根据公告id找到一条公告
     public List<Notice> findAll(int dormitory_id); //根据宿舍楼号返回所有公告
-    public int sum(int dormitory_id);   //根据宿舍楼号返回所有公告数量
+    public long sum(int dormitory_id);   //根据宿舍楼号返回所有公告数量
 
 }
