+++ src/dao/iml/UserDaoIml.java
@@ -73,11 +73,11 @@ public class UserDaoIml implements UserDao {
     }
 
     @Override
-    public int sum(int dormitory_id) {
+    public long sum(int dormitory_id) {
         try{
             QueryRunner runner = new QueryRunner(JdbcUtils.getDataSource());
             String sql = "select count(*) from user where dormitory_id=?";
-            return (int) runner.query(sql, dormitory_id, new ScalarHandler());
+            return (long) runner.query(sql, dormitory_id, new ScalarHandler());
         } catch(Exception e){
             throw new RuntimeException(e);
         }
